build basic navbar component, refactor app component
