+++ src/App.js
@@ -1,18 +1,13 @@
 import React, { useState } from "react";
 import "./App.scss";
 import "./styles/base.scss";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link,
-  useRouteMatch,
-} from "react-router-dom";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 import NotFound from "./components/NotFound";
 import AddItem from "./components/AddItem";
 import ListContainer from "./components/ShoppingList/ListContainer";
 import Home from "./components/Home";
+import Navbar from "./components/Navbar";
 
 const App = () => {
   const [auth, setAuth] = useState(localStorage.getItem("token"));
@@ -21,8 +16,7 @@ const App = () => {
     <Router>
       {auth ? (
         <div className="App">
-          <MenuLink activeWhenExact={true} pathTo="/List" label="List View" />
-          <MenuLink pathTo="/AddItem" label="Add Item" />
+          <Navbar />
           <Switch>
             <Route exact path="/">
               <ListContainer />
@@ -40,19 +34,6 @@ const App = () => {
         <Home setAuth={setAuth} />
       )}
     </Router>
-  );
-};
-
-const MenuLink = ({ label, pathTo, activeWhenExact = false }) => {
-  let match = useRouteMatch({
-    path: pathTo,
-    exact: activeWhenExact,
-  });
-
-  return (
-    <div className={match ? "active" : ""}>
-      <Link to={pathTo}>{label}</Link>
-    </div>
   );
 };
